changes.py: suspect confidence intervals for UK / England
I "suspect" the confidence intervals for England should not be the sum
of the confidence intervals for the regions of England.

I think you would only add them up if the error was totally correlated.
The error in this case is sampling error, so it's supposed to be fully
random?