jump.txt: remove bad ?? entry
I don't remember what this was.  It wasn't an old entry from jump.ods,
I can't see what it could possibly refer to.  And I had specifically
reviewed jumps in incidence_history.  The two recent jumps in
incidence_hisory but not incidence turned out to be due to the anomaly
noted on 08/07/22.